+++ packages/react-native-contentpass-cmp-consentmanager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @contentpass/react-native-contentpass-cmp-consentmanager
 
+## 0.1.1
+
+### Patch Changes
+
+- Improve OneTrust integration and handling of first layer links
+
 ## 0.1.0
 
 ### Minor Changes